Removed error from theme.js
I wrapped the event listener in an if statement to first check if there actually is a theme button to put the event listener on before applying it so that an error does not occur on secondary pages that do not contain the button

+++ js/theme.js
@@ -42,21 +42,24 @@ window.addEventListener('load', (e) => {
     }
 });
 
-window.addEventListener("DOMContentLoaded", () => {
-    const themeSwitch = document.querySelector('.theme-switch-container');
-    const body = document.querySelector('body');
-    themeSwitch.addEventListener('click', () => {
-        if (body.classList.contains('light')) {
-            localStorage.clear();
-            localStorage.setItem('theme', 'dark');
-            setDarkMode();
-        } else if (body.classList.contains('dark')) {
-            localStorage.clear();
-            localStorage.setItem('theme', 'light');
-            setLightMode();
-        }
+if (document.getElementById("theme-switch") != null) {
+    window.addEventListener("DOMContentLoaded", () => {
+        const themeSwitch = document.querySelector('.theme-switch-container');
+        const body = document.querySelector('body');
+        themeSwitch.addEventListener('click', () => {
+            if (body.classList.contains('light')) {
+                localStorage.clear();
+                localStorage.setItem('theme', 'dark');
+                setDarkMode();
+            } else if (body.classList.contains('dark')) {
+                localStorage.clear();
+                localStorage.setItem('theme', 'light');
+                setLightMode();
+            }
+        });
     });
-});
+}
+
